refactor: move getDisplayParty into ProjectDetails.utils.tsx
