+++ web-registry/src/components/templates/ProjectDetails/ProjectDetails.utils.tsx
@@ -1,14 +1,21 @@
 import StaticMap from 'web-components/lib/components/map/StaticMap';
 import { GalleryPhoto } from 'web-components/lib/components/organisms/Gallery/Gallery.types';
 import { Asset } from 'web-components/lib/components/sliders/ProjectMedia';
+import { Party } from 'web-components/lib/components/user/UserInfo';
 
-import { Document, Maybe, Project } from 'generated/graphql';
+import {
+  Document,
+  Maybe,
+  PartyFieldsFragment,
+  Project,
+} from 'generated/graphql';
 import { AllCreditClassQuery } from 'generated/sanity-graphql';
 import {
   AnchoredProjectMetadataBaseLD,
   LegacyProjectMetadataLD,
   NameImageDescription,
   ProjectPageMetadataLD,
+  ProjectStakeholder,
 } from 'lib/db/types/json-ld';
 
 import {
@@ -131,3 +138,61 @@ export const getProjectGalleryPhotos = ({
 
   return photos;
 };
+
+export function getParty(
+  party?: Maybe<PartyFieldsFragment>,
+): Party | undefined {
+  if (!party) {
+    return undefined;
+  }
+
+  return {
+    name: party.name,
+    description: party.description,
+    type: party.type,
+    image: party.image,
+    address: party.walletByWalletId?.addr || '',
+    link: party?.websiteLink,
+  };
+}
+
+type StakeholderType =
+  | 'regen:projectDeveloper'
+  | 'regen:landSteward'
+  | 'regen:landOwner'
+  | 'regen:projectOriginator';
+
+const getPartyFromMetadata = (
+  metadata: AnchoredProjectMetadataBaseLD,
+  role: StakeholderType,
+): Party | undefined => {
+  const metadataRole: ProjectStakeholder | undefined = metadata[role];
+  if (!metadataRole) return undefined;
+
+  return {
+    name: metadataRole?.['schema:name'] || '',
+    description: metadataRole?.['schema:description'] || '',
+    type: metadataRole?.['@type'].includes('regen:Organization') // covers Organization or OrganizationDisplay
+      ? 'ORGANIZATION' // to provide default image
+      : 'USER',
+    image: metadataRole?.['schema:image'],
+    location: metadataRole?.['schema:location']?.place_name || '',
+    address: metadataRole?.['regen:adress'] || '',
+    link: metadataRole?.['schema:url'] || '',
+  };
+};
+
+export function getDisplayParty(
+  role: StakeholderType,
+  metadata?: AnchoredProjectMetadataBaseLD,
+  party?: Maybe<PartyFieldsFragment>,
+): Party | undefined {
+  const showOnProjectPage = metadata?.[role]?.['regen:showOnProjectPage'];
+  if (showOnProjectPage) {
+    const dbParty = getParty(party);
+    if (dbParty) return dbParty;
+    // If no party info available for this role, check the metadata
+    return getPartyFromMetadata(metadata, role);
+  }
+  return undefined;
+}
